REFACTOR: Add additional login logic
Added one more logic as below.

On signup page, if the user trying to sign up with an id which already
exists in the database the browser will alert the user saying "This ID
already exists! Try another ID". And if it doesn't exist, it'll work as
same as the former commit.