REFACTOR: Edit 'accept' function
Edited the 'accept' function. It was deleting the original one and save
the new one only changing the state. But from the server, the save
function gets the info and use Optional function to check whether the
object exists in the database. If so, it edits the state of it, if not,
it just saves the info as new one.